+++ emcc.py
@@ -2002,6 +2002,23 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
     if not shared.Settings.LEGALIZE_JS_FFI:
       assert shared.Building.is_wasm_only(), 'LEGALIZE_JS_FFI incompatible with RUNNING_JS_OPTS.'
 
+    # check if we can address the 2GB mark and higher: either if we start at
+    # 2GB, or if we allow growth to either any amount or to 2GB or more.
+    if shared.Settings.WASM_BACKEND and \
+       (shared.Settings.INITIAL_MEMORY > 2 * 1024 * 1024 * 1024 or
+        (shared.Settings.ALLOW_MEMORY_GROWTH and
+         (shared.Settings.MAXIMUM_MEMORY < 0 or
+          shared.Settings.MAXIMUM_MEMORY > 2 * 1024 * 1024 * 1024))):
+      shared.Settings.CAN_ADDRESS_2GB = 1
+      if shared.Settings.MALLOC == 'emmalloc':
+        if shared.Settings.INITIAL_MEMORY >= 2 * 1024 * 1024 * 1024:
+          suggestion = 'decrease INITIAL_MEMORY'
+        elif shared.Settings.MAXIMUM_MEMORY < 0:
+          suggestion = 'set MAXIMUM_MEMORY'
+        else:
+          suggestion = 'decrease MAXIMUM_MEMORY'
+        exit_with_error('emmalloc only works on <2GB of memory. Use the default allocator, or ' + suggestion)
+
     shared.Settings.EMSCRIPTEN_VERSION = shared.EMSCRIPTEN_VERSION
     shared.Settings.PROFILING_FUNCS = options.profiling_funcs
     shared.Settings.SOURCE_MAP_BASE = options.source_map_base or ''
@@ -3179,6 +3196,8 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
       cmd += ['--table-max=-1']
     if shared.Settings.SIDE_MODULE:
       cmd += ['--mem-max=-1']
+    elif not shared.Settings.ALLOW_MEMORY_GROWTH:
+      cmd += ['--mem-max=' + str(shared.Settings.INITIAL_MEMORY)]
     elif shared.Settings.MAXIMUM_MEMORY >= 0:
       cmd += ['--mem-max=' + str(shared.Settings.MAXIMUM_MEMORY)]
     if shared.Settings.LEGALIZE_JS_FFI != 1:
@@ -3268,6 +3287,12 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
   # pthreads memory growth requires some additional JS fixups
   if shared.Settings.USE_PTHREADS and shared.Settings.ALLOW_MEMORY_GROWTH:
     final = shared.Building.apply_wasm_memory_growth(final)
+
+  # >=2GB heap support requires pointers in JS to be unsigned. rather than
+  # require all pointers to be unsigned by default, which increases code size
+  # a little, keep them signed, and just unsign them here if we need that.
+  if shared.Settings.CAN_ADDRESS_2GB:
+    final = shared.Building.use_unsigned_pointers_in_js(final)
 
   if shared.Settings.OPT_LEVEL >= 2 and shared.Settings.DEBUG_LEVEL <= 2:
     # minify the JS