+++ src/settings_internal.js
@@ -173,6 +173,10 @@ var WASM_FUNCTIONS_THAT_ARE_NOT_NAME_MANGLED = ['setTempRet0', 'getTempRet0', 's
 // Internal: value of -flto argument (either full or thin)
 var LTO = 0;
 
+// Whether we may be accessing the address 2GB or higher. If so then we need
+// to be using unsigned pointers in JS.
+var CAN_ADDRESS_2GB = 0;
+
 // Whether to emit DWARF in a separate wasm file on the side (this is not called
 // "split" because there is already a DWARF concept by that name).
 // When DWARF is on the side, the main file has no DWARF info, while the side